+++ pkg/validator/hocon.go
@@ -8,12 +8,18 @@ import (
 // HOCON file.
 type HoconValidator struct{}
 
+var _ Validator = HoconValidator{}
+
 // Validate checks if the provided byte slice represents a valid .hocon file.
-func (HoconValidator) Validate(b []byte) (bool, error) {
+func (HoconValidator) ValidateSyntax(b []byte) (bool, error) {
 	_, err := hocon.ParseString(string(b))
 	if err != nil {
 		return false, err
 	}
 
 	return true, nil
+}
+
+func (HoconValidator) ValidateFormat(_ []byte, _ any) (bool, error) {
+	return false, ErrMethodUnimplemented
 }